[MYTASKS-7]: Organize folders (task-service)
